refactor: 쿼리 수정
