+++ src/main/java/com/example/sparta_3rd_newsfeed/friend/repository/FriendRepository.java
@@ -1,11 +1,12 @@
 package com.example.sparta_3rd_newsfeed.friend.repository;
 
-import com.example.sparta_3rd_newsfeed.Member;
+import com.example.sparta_3rd_newsfeed.member.entity.Member;
 import com.example.sparta_3rd_newsfeed.friend.entity.Friend;
+import com.example.sparta_3rd_newsfeed.friend.entity.FriendStatus;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -18,19 +19,16 @@ public interface FriendRepository extends JpaRepository<Friend, Long> {
     }
     Optional<Friend> findBySenderAndReceiver(Member sender, Member receiver);
 
-    @Query("SELECT f FROM Friend f JOIN FETCH f.sender JOIN FETCH f.receiver WHERE f.receiver.id = :receiverId AND f.status = 'ACCEPTED'")
-    Page<Friend> findByReceiver(Long receiverId, Pageable pageable);
+    @EntityGraph(attributePaths = {"sender"})
+    Page<Friend> findByReceiverId(Long receiverId, FriendStatus status, Pageable pageable);
 
-    @Query("SELECT f FROM Friend f JOIN FETCH f.sender JOIN FETCH f.receiver WHERE f.receiver.id = :receiverId AND f.status = 'ACCEPTED' AND f.sender.memberName LIKE %:name%")
-    Page<Friend> findByReceiverAndName(Long receiverId, String name, Pageable pageable);
+    @EntityGraph(attributePaths = {"sender"})
+    Page<Friend> findByReceiverIdAndName(Long receiverId, FriendStatus status, String name, Pageable pageable);
 
-    @Query("SELECT f FROM Friend f JOIN FETCH f.sender JOIN FETCH f.receiver WHERE f.sender.id = :senderId AND f.status = 'ACCEPTED'")
-    Page<Friend> findBySender(Long senderId, Pageable pageable);
+    @EntityGraph(attributePaths = {"receiver"})
+    Page<Friend> findBySenderId(Long senderId, FriendStatus status, Pageable pageable);
 
-    @Query("SELECT f FROM Friend f JOIN FETCH f.sender JOIN FETCH f.receiver WHERE f.sender.id = :senderId AND f.status = 'ACCEPTED' AND f.receiver.memberName LIKE %:name%")
-    Page<Friend> findBySenderAndName(Long senderId, String name, Pageable pageable);
-
-    @Query("SELECT f FROM Friend f JOIN FETCH f.sender JOIN FETCH f.receiver WHERE f.receiver.id = :memberId AND f.status = 'PENDING'")
-    Page<Friend> findPendingRequests(Long memberId, Pageable pageable);
+    @EntityGraph(attributePaths = {"receiver"})
+    Page<Friend> findBySenderIdAndName(Long senderId, FriendStatus status, String name, Pageable pageable);
 
 }
